Update README, move metrics to DCGAN folder
